refactor(authentication): JpaRepository替换为BaseJpaRepository，替换JpaRepository中的反模式方法
